added useRecipes.ts exported functions (Custom Hooks) added name to Name and description to Description on web site.
