+++ frontend/src/Recipe.ts
@@ -1,7 +1,9 @@
 export type Recipe = {
     id: string,
     name: string,
+    description: string,
 }
 export type NewRecipe = {
-    name: string
+    name: string,
+    description: string,
 }
